Consolidate pings/maker.ts into one sync file
